restructured components and folders
