+++ src/App.js
@@ -1,7 +1,7 @@
 // import { Route, Routes } from 'react-router';
 import './App.css';
-import Footer from './components/CommonComponents/Footer.js';
-import Nav from './components/CommonComponents/Nav.js';
+import Footer from './components/Footer.js';
+import Nav from './components/Nav.js';
 
 function App() {
   return (
