+++ src/components/Nav.js
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import logo from "../../assets/logo_black.png"
+import logo from "../assets/logo_black.png"
 
 const Nav = () => {
     return (
